Move file into dedicated folder organization
